django/requirements.txt
Django>=2.1.5,<3
python-decouple>=3.1,<4
djangorestframework>=3.9.1,<4
django-filter>=2.1.0,<3
markdown>=3.0.1,<4
django-cors-headers>=2.4.0,<3
psycopg2-binary>=2.7.7,<3